tests/validation/python/requirements.txt
# VULNERABLE: requests 2.25.0 - CVE-2023-32681
# Safe version: 2.31.0+
requests==2.25.0

# VULNERABLE: pyyaml 5.3 - CVE-2020-14343
# Safe version: 5.4+
pyyaml==5.3

# VULNERABLE: pillow 8.0.0 - Multiple CVEs
# Safe version: 9.3.0+
# This one is NOT used in code - should be unreachable
pillow==8.0.0

# Safe for comparison
click==8.1.7
